# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "My Backend Project API",
        "description": "Application API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
...
        "/profile": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Profile"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Not authorized"
                    }
                },
                "tags": [
                    "user"
                ],
                "operationId": "profile"
            }
        }
    },
    "definitions": {
...
        "Profile": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string",
                    "example": "somelogin",
                    "required": true
                },
                "name": {
                    "type": "string",
                    "example": "somename",
                    "required": true
                },
                "surname": {
                    "type": "string",
                    "example": "somesurname"
                },
                "gender": {
                    "type": "string",
                    "example": "male/female"
                },
                "age": {
                    "type": "number",
                    "example": 42
                },
                "postCount": {
                    "type": "number",
                    "required": true
                },
                "subCount": {
                    "type": "number",
                    "required": true
                }
            }
        }
    },
    "schemes": [
        "http"
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ]
}